+++ DuDoong-Batch/src/main/java/band/gosrock/job/EventSettlementPDF.java
@@ -1,0 +1,65 @@
+package band.gosrock.job;
+
+
+import band.gosrock.domain.domains.event.adaptor.EventAdaptor;
+import band.gosrock.infrastructure.config.pdf.PdfRender;
+import band.gosrock.parameter.EventJobParameter;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
+import org.springframework.batch.core.configuration.annotation.JobScope;
+import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.thymeleaf.context.Context;
+import org.thymeleaf.spring5.SpringTemplateEngine;
+
+@Slf4j
+@RequiredArgsConstructor
+@Configuration
+public class EventSettlementPDF {
+
+    private static final String JOB_NAME = "이벤트정산서";
+    private static final String BEAN_PREFIX = JOB_NAME + "_";
+
+    private final JobBuilderFactory jobBuilderFactory;
+    private final StepBuilderFactory stepBuilderFactory;
+    private final EventAdaptor eventAdaptor;
+
+    @Qualifier(BEAN_PREFIX + "eventJobParameter")
+    private final PdfRender pdfRender;
+
+    private final SpringTemplateEngine templateEngine;
+
+    @Bean(BEAN_PREFIX + "eventJobParameter")
+    @JobScope
+    public EventJobParameter eventJobParameter() {
+        return new EventJobParameter(eventAdaptor);
+    }
+
+    @Bean(JOB_NAME)
+    public Job slackUserStatisticJob() {
+        return jobBuilderFactory.get(JOB_NAME).preventRestart().start(userStatisticStep()).build();
+    }
+
+    @Bean(BEAN_PREFIX + "step")
+    @JobScope
+    public Step userStatisticStep() {
+        return stepBuilderFactory
+                .get(BEAN_PREFIX + "step")
+                .tasklet(
+                        (contribution, chunkContext) -> {
+                            Context context = new Context();
+                            context.setVariable("username", "우저이름");
+                            String signUp = templateEngine.process("settlement", context);
+                            pdfRender.generatePdfFromHtml(signUp);
+
+                            return RepeatStatus.FINISHED;
+                        })
+                .build();
+    }
+}
